# Conflicts:
#	backend/src/api/employee.controller.ts
#	backend/src/api/position.controller.ts
#	backend/src/api/referral.controller.ts
#	db/src/db.module.ts
#	db/src/employee/employee.controller.ts
#	db/src/employee/employee.service.ts
#	db/src/entities/Position.ts
#	db/src/main.ts
#	db/src/position/position.controller.ts
#	db/src/position/position.module.ts
#	db/src/position/position.service.ts
#	db/src/referral/referral.controller.ts
#	db/src/referral/referral.dto.ts
#	db/src/referral/referral.module.ts
#	db/src/referral/referral.service.ts
#	frontend/src/App.js
#	frontend/src/components/Footer/Footer.module.css
#	frontend/src/components/Header/Header.js
#	frontend/src/components/Header/Header.module.css
#	frontend/src/components/Main/Main.js
#	frontend/src/components/Main_Home/main_home.js
#	frontend/src/components/Main_Mailbox/email_bar.js
#	frontend/src/components/Main_Mailbox/main_mailbox.js
#	frontend/src/components/Main_Previous_Ref/Main_previous_ref.js
#	frontend/src/components/NavBar/navBar.js
#	frontend/src/components/NavBar/navBar.module.css

+++ db/src/entities/Position.ts
@@ -15,12 +15,11 @@ import { Referral } from './Referral';
 @Entity('position', { schema: 'aki' })
 export class Position {
   // Matt Cappucci - made this auto increment
-  @PrimaryGeneratedColumn({ type: 'bigint', name: 'id' })
+  @PrimaryGeneratedColumn({ type: 'int', name: 'id' })
   id: number;
 
   /*@Column("bigint", { primary: true, name: "id" })
   id: string;*/
-
   @Column('longtext', { name: 'description', nullable: true })
   description: string | null;
